Feat/develop lawyer service ver 1.0.0

+++ rating-service/src/main/java/org/binitshrestha/ratingservice/dto/RatingCreateRequest.java
@@ -7,9 +7,8 @@ public record RatingCreateRequest(
     Long userId,
     Long lawyerId,
     Long caseId,
-
     String ratingType,
-    int ratingValue,
+    double ratingValue,
     String comments
 ) {
 }
